Restructured Tests dir to reflect to group together the basic functionalities of the library: NN, Lin Alg, Tensor Ops, Tensor Interations (Tensor), and IO
