--- Tests/NeuralNet/NeuralNetTest.go
@@ -1,1 +1,0 @@
-package TG
